copy tests from TestSorted class
